dictapp/requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
gunicorn==20.1.0
idna==2.10
lxml==4.6.3
python-decouple==3.4
pytz==2021.1
requests==2.25.1
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
